baza lokalna

+++ TeamOrganizer/MySqlInitializer.cs
@@ -23,15 +23,15 @@ namespace TeamOrganizer
                 var migrationHistoryTableExists = ((IObjectContextAdapter)context).ObjectContext.ExecuteStoreQuery<int>(
                 string.Format(
                   "SELECT COUNT(*) FROM information_schema.tables WHERE table_schema = '{0}' AND table_name = '__MigrationHistory'",
-                  "[Insert your database schema here - such as `users`]"));
+                  "[`ziwg`]"));
 
                 // if MigrationHistory table is not there (which is the case first time we run) - create it
-              /*  if (migrationHistoryTableExists.FirstOrDefault() == 0)
+                if (migrationHistoryTableExists.FirstOrDefault() == 0)
                 {
                     context.Database.Delete();
                     context.Database.Create();
                 }
-                */
+                
             }
         }
     }
